Make regexp.h internal, only visible to the new tfprof contrib lib.
RE2 is not as portable as we would like, so we want to try to discourage
the use of it in core libraries.  After having removed it from
all core uses, the only remaining use is in tfprof, so we move
its visibility to only be to tfprof, which is an external tool.

tfprof appears to fundamentally rely on regexp as a user interface, so
it cannot be easily removed.  However, perhaps one day tfprof will
move into its own repo and then we can drop RE2 from our builds
entirely.
Change: 134120532