Don't pass serialized tf.Example column to _get_predict_ops().
The dict passed to that function should have the same schema at training and serving time.  The fact that serving_input_fn() uses the features dict as a means to pass the unparsed tf.Example column out is expedient but weird (and it just bit a user).  Properly factoring that apart (eventually) would be an API-breaking change, but for now we can at least ensure that the unparsed column is not passed to the model (which cannot possibly have a use for it).
Change: 135711412