Mutes warnings for some deprecated arguments with default values.  Fixes a bug.
Users were getting log warnings even when using the API correctly,
because of argument passthrough within tf.learn.  This change mutes
warnings for calls to fit, estimate, and predict if the deprecated
arguments receive default values.

Also fixes a bug in the deprecation tool where "==" was used instead of "is" to compare objects, which was at times yeilding a non-boolean.
Change: 142162193