Make legacy_seq2seq seq2seq models perform a shallow copy of the incoming cell.
this forces a separate cell instance to be used in the encoder and decoder(s).
right now this is a no-op, but most RNNCells will soon keep track of their
variables and using the same cell for an encoder and decoder pair without
tied weights will not be legal.  at that point, this change will still
be a no-op, but without it the RNNCells would raise an error.
Change: 146423203