Avoid merging adjacent XLA compilations from different scopes/functions
This is part 2 of the bugfix.  It implements the xla chain breaking mechanism
based on different coloring of the graph (as represented by different XlaScope
strings).  In part 1, we modified both experimental_jit_scope and Defun to
mark their ops as having different XlaScopes, so this is the final change that
actually enables the fusion breaking.

Also fixed a bug where xla_enabled was not True for rnn_cell_tests (the XLA benchmarks in that test were not actually being run with xla enabled).
Change: 148286731