Add new ops for memory statistics.
This CL adds only the MaxBytesInUse op, which collects the peak memory usage of
a device allocator. Other ops can be added similarly when demanded. For now, we
only enable MaxBytesInUse for GPU because memory statistics are unreliable for
CPU allocators.

This CL essentially merges part of Yaroslav Bulatov's work on
https://github.com/yaroslavvb/memory_probe_ops to TensorFlow.
Change: 148854571