Adds new Dataset APIs for filesystems
This change adds a new API to construct a Dataset:
 - Dataset.list_files("/path/to/files/*.data")

This API can be used in conjunction with the forthcoming .interleave()
API which will allow for the construction of parallel loading of data files.

As an added bonus, input pipelines constructed in this manner should be
deterministic, to allow for reproducible work.

PiperOrigin-RevId: 159611145