Allow ffmpeg `samples_per_second` to be a Tensor
This changes the `samples_per_second` parameter of the `encode_audio`
and `decode_audio` ops from an `Attr` to an `Input`, so that it can be
given arbitrary tensor values instead of only constants.

This change is important for use cases that want to use a single graph
to encode audio clips at arbitrary sample rates. (In particular, we want
to create a Python function that uses a long-running TensorFlow session
to encode audio; the sample rate cannot be known ahead of time, and we
don't want to have to reconstruct the graph on every call.)

PiperOrigin-RevId: 164799067